code updates related to Refresh Exchange GUI
