+++ Product/Production/Adapters/General/CONNECTAdminGUI/src/main/webapp/emDialogRefreshExchange.xhtml
@@ -1,0 +1,33 @@
+<ui:composition xmlns:h="http://java.sun.com/jsf/html" xmlns:ui="http://java.sun.com/jsf/facelets" xmlns:p="http://primefaces.org/ui" xmlns="http://www.w3.org/1999/xhtml"
+                xmlns:f="http://java.sun.com/jsf/core">
+
+    <div id="emRefreshDiv">
+        <p:dialog id="refreshExchangeStatusDlg" header="Exchange Refresh Status" widgetVar="wvDlgRefreshExchangeStatus" modal="true" dynamic="true" draggable="false" closeOnEscape="true" resizable="false" width="650" height="auto">
+            <div class="content-block content">
+                <h:form id="refreshExchangeStatusForm" class="form-horizontal" role="form">
+                    <p:dataList id="dataTableExchangeDownload" var="exStatus" value="#{exchangeManagerBean.exDownloadStatus}"
+                                 type="unordered" itemType="none" paginator="false">
+                        <p:column>
+                            <p:commandLink styleClass="glyphicon glyphicon-remove-circle highlightRED pull-left" rendered="#{exStatus.success eq 'false'}"/>
+                            <p:commandLink styleClass="glyphicon glyphicon-ok-circle highlightGREEN pull-left" rendered="#{exStatus.success eq 'true'}"/>
+                            <h:outputText value="&#160;" />
+                            <h:outputText value="#{exStatus.name}" style="display:inline-block"/>
+                        </p:column>
+                        <p:column>
+                            <p:dataList value="#{exStatus.stepStatus}" var="step" type="unordered" itemType="none" paginator="false" style="border:none">
+                                <h:outputText value="#{step.message}"/>
+                            </p:dataList>
+                        </p:column>
+                    </p:dataList>
+                </h:form>
+                <div class="form-group">
+                   <div class="col-sm-12">
+                        <div class="form-button-row">
+                          <p:commandButton id="closeRefreshExchange" styleClass="btn btn-primary pull-right" value="Close" onclick="PF('wvDlgRefreshExchangeStatus').hide()"  update=":tabviewExchange"/>
+                        </div>
+                   </div>
+                </div>
+            </div>
+        </p:dialog>
+    </div>
+</ui:composition>
