+++ Product/Production/Adapters/General/CONNECTAdminGUI/src/main/webapp/exchangeManager.xhtml
@@ -106,7 +106,7 @@
                                                         <p:commandButton styleClass="btn refresh-button"
                                                             icon="ui-icon-refresh"
                                                             actionListener="#{exchangeManagerBean.refreshExchangeInfo}"
-                                                            value="Refresh Exchanges" update=":tabviewExchange" />
+                                                            value="Refresh Exchanges" update=":tabviewExchange :refreshExchangeStatusDlg" />
                                                     </div>
                                                 </div>
                                             </div>
@@ -146,7 +146,7 @@
                                                         <h:outputText rendered="#{not rec.disabled}" value="Yes" />
                                                     </p:column>
                                                     <p:column headerText="Last Updated">
-                                                        <h:outputText value="#{rec.lastUpdated}">
+                                                        <h:outputText value="#{rec.lastUpdated.toGregorianCalendar().time}">
                                                             <f:convertDateTime pattern="MM-dd-yyyy HH:mm:ss" />
                                                         </h:outputText>
                                                     </p:column>
@@ -317,6 +317,7 @@
             </p:confirmDialog>
             <p:growl id="msgGrowl" for="msgForGrowl" closable="true" />
             <ui:include src="emDialogSaveExchange.xhtml" />
+            <ui:include src="emDialogRefreshExchange.xhtml" />
         </ui:define>
     </ui:composition>
 </h:body>
