+++ Product/Production/Common/CONNECTCommonWeb/pom.xml
@@ -40,7 +40,6 @@
             <groupId>org.connectopensource</groupId>
             <artifactId>ConnectionManagerCore</artifactId>
             <version>${project.parent.version}</version>
-            <scope>runtime</scope>
         </dependency>
         <dependency>
             <groupId>org.connectopensource</groupId>
