--- Product/Production/Common/CONNECTCommonWeb/src/test/java/gov/hhs/fha/nhinc/exchangemgr/ExchangeScheduledTaskTest.java
@@ -1,87 +1,0 @@
-/*
- * Copyright (c) 2009-2017, United States Government, as represented by the Secretary of Health and Human Services.
- * All rights reserved.
- *
- * Redistribution and use in source and binary forms, with or without
- * modification, are permitted provided that the following conditions are met:
- *     * Redistributions of source code must retain the above
- *       copyright notice, this list of conditions and the following disclaimer.
- *     * Redistributions in binary form must reproduce the above copyright
- *       notice, this list of conditions and the following disclaimer in the documentation
- *       and/or other materials provided with the distribution.
- *     * Neither the name of the United States Government nor the
- *       names of its contributors may be used to endorse or promote products
- *       derived from this software without specific prior written permission.
- *
- * THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS "AS IS" AND
- * ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE IMPLIED
- * WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE ARE
- * DISCLAIMED. IN NO EVENT SHALL THE UNITED STATES GOVERNMENT BE LIABLE FOR ANY
- * DIRECT, INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES
- * (INCLUDING, BUT NOT LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES;
- * LOSS OF USE, DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND
- * ON ANY THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT
- * (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS
- * SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
- */
-package gov.hhs.fha.nhinc.exchangemgr;
-
-import gov.hhs.fha.nhinc.connectmgr.persistance.dao.ExchangeInfoDAOFileImpl;
-import gov.hhs.fha.nhinc.connectmgr.uddi.UDDIAccessorException;
-import gov.hhs.fha.nhinc.connectmgr.uddi.UDDIUpdateManagerHelper;
-import java.io.File;
-import java.net.URISyntaxException;
-import java.net.URL;
-import javax.xml.datatype.DatatypeConfigurationException;
-import static org.junit.Assert.assertNotNull;
-import static org.junit.Assert.assertTrue;
-import org.junit.Test;
-import static org.mockito.Matchers.anyString;
-import static org.mockito.Mockito.mock;
-import static org.mockito.Mockito.when;
-import org.uddi.api_v3.BusinessDetail;
-
-/**
- *
- * @author tjafri
- */
-public class ExchangeScheduledTaskTest {
-
-    private static final String FILE_NAME = "/config/ExchangeScheduledTaskTest/exchangeInfoTest.xml";
-
-    @Test
-    public void task() throws ExchangeManagerException, UDDIAccessorException, DatatypeConfigurationException {
-        final ExchangeInfoDAOFileImpl exDao = createExchangeInfoDAO(FILE_NAME);
-        final UDDIUpdateManagerHelper uddiMgr = mock(UDDIUpdateManagerHelper.class);
-        ExchangeScheduledTask workerThread = new ExchangeScheduledTask() {
-            @Override
-            protected ExchangeInfoDAOFileImpl getExchangeDAO() {
-                return exDao;
-            }
-
-            @Override
-            protected UDDIUpdateManagerHelper getUDDIManager() {
-                return uddiMgr;
-            }
-        };
-
-        when(uddiMgr.forceRefreshUDDIFile(anyString())).thenReturn(new BusinessDetail());
-        workerThread.task();
-        assertNotNull(exDao.loadExchangeInfo().getExchanges().getExchange().get(0).getLastUpdated());
-    }
-
-    private ExchangeInfoDAOFileImpl createExchangeInfoDAO(String filename) {
-        try {
-            URL url = this.getClass().getResource(filename);
-            File exchangeInfoFile = new File(url.toURI());
-            assertTrue("File does not exist: " + exchangeInfoFile, exchangeInfoFile.exists());
-            ExchangeInfoDAOFileImpl exchangeDAO = ExchangeInfoDAOFileImpl.getInstance();
-            exchangeDAO.setFileName(exchangeInfoFile.getAbsolutePath());
-
-            return exchangeDAO;
-        } catch (URISyntaxException e) {
-            throw new RuntimeException(e);
-        }
-    }
-
-}
