--- Product/Production/Common/CONNECTCommonWeb/src/test/resources/config/ExchangeScheduledTaskTest/exchangeInfoTest.xml
@@ -1,14 +1,0 @@
-<?xml version="1.0" encoding="UTF-8" standalone="yes"?>
-<exchangeInfo xmlns="urn:gov:hhs:fha:nhinc:exchange" xmlns:ns2="urn:gov:hhs:fha:nhinc:exchange:directory" >
-    <refreshInterval>1440</refreshInterval>
-    <maxNumberOfBackups>1</maxNumberOfBackups>
-    <refreshActive>true</refreshActive>
-    <exchanges>
-        <exchange type="uddi">
-            <name>Test UDDI Exchnage</name>
-            <url>http://localhost/uddi/inquiry</url>
-            <organizationList>
-            </organizationList>
-        </exchange>
-    </exchanges>
-</exchangeInfo>
