+++ Product/Production/Common/CONNECTCoreLib/src/main/java/gov/hhs/fha/nhinc/nhinclib/NhincConstants.java
@@ -184,6 +184,7 @@ public class NhincConstants {
     public static final String XDS_HOME_COMMUNITY_ID_PROPERTY = "XDSbHomeCommunityId";
     public static final String MAX_UDDI_RESULTS_PROPERTY = "UDDIMaxResults";
     public static final String FHIR_DIRECTORY_FILE = "FHIRDirectoryMapping";
+    public static final String MESSAGES_PROPERTY_FILE = "messages";
     // Concurrent Executor Service Constants (used to retrieve values from
     // gateway.properties)
     public static final String CONCURRENT_POOL_SIZE = "ConcurrentPoolSize";
@@ -583,6 +584,7 @@ public class NhincConstants {
     public static final String ADMIN_CERT_LIST_KEYSTORE = "listKeyStores";
     public static final String ADMIN_CERT_DELETE = "deleteCertificate";
     public static final String ADMIN_CERT_EDIT = "editCertificate";
+
     private NhincConstants() {
     }
 }
