+++ Product/Production/Common/Properties/src/main/resources/messages.properties
@@ -1,0 +1,10 @@
+# To change this license header, choose License Headers in Project Properties.
+# To change this template file, choose Tools | Templates
+# and open the template in the editor.
+DOWNLOAD_SUCCESS=Download successful.
+DOWNLOAD_FAILURE=Download failed.
+TRANSFORM_SUCCESS=Exchange refresh successful.
+TRANSFORM_FAILURE=Exchange refresh failed.
+SCHEMA_VAL_SUCCESS=Schema Validation successful.
+SCHEMA_VAL_FAILED=Schema Validation failed.
+SCHEMA_VAL_SKIP=Schema Validation skipped.
