+++ Product/Production/Services/ConnectionManagerCore/src/main/java/gov/hhs/fha/nhinc/exchangemgr/util/ExchangeManagerUtil.java
@@ -24,37 +24,26 @@
  * (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS
  * SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
-package gov.hhs.fha.nhinc.admingui.services;
+package gov.hhs.fha.nhinc.exchangemgr.util;
 
-import gov.hhs.fha.nhinc.admingui.model.ConnectionEndpoint;
-import gov.hhs.fha.nhinc.exchange.ExchangeInfoType;
-import gov.hhs.fha.nhinc.exchange.ExchangeType;
-import gov.hhs.fha.nhinc.exchange.directory.OrganizationType;
-import gov.hhs.fha.nhinc.exchangemgr.util.ExchangeDownloadStatus;
+import java.util.ArrayList;
 import java.util.List;
 
 /**
- * @author tran tang
  *
+ * @author tjafri
  */
-public interface ExchangeManagerService {
+public class ExchangeManagerUtil {
 
-    public boolean saveExchange(ExchangeType exchange);
+    private final ExchangeDateUpdateMgr exScheduleTask = new ExchangeDateUpdateMgr();
+    List<ExchangeDownloadStatus> downloadStatus;
 
-    public boolean deleteExchange(String exchangeName);
+    public List<ExchangeDownloadStatus> getDownloadStatus() {
+        return downloadStatus;
+    }
 
-    public List<ExchangeType> getAllExchanges();
-
-    public List<OrganizationType> getAllOrganizations(String exchangeName);
-
-    public List<ConnectionEndpoint> getAllConnectionEndpoints(String exchangeName, String hcid);
-
-    public ExchangeInfoType getExchangeInfoView();
-
-    public boolean saveGeneralSetting(ExchangeInfoType exchangeInfo);
-
-    public List<ExchangeDownloadStatus> refreshExchangeManager();
-
-    public boolean pingService(ConnectionEndpoint connEndpoint);
-
+    public List<ExchangeDownloadStatus> forceRefreshExchanges() {
+        downloadStatus = new ArrayList<>();
+        return exScheduleTask.task(downloadStatus);
+    }
 }
